backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database - PostgreSQL
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Database - MongoDB
pymongo==4.6.0
motor==3.3.2

# Redis
redis==5.0.1
hiredis==2.2.3

# Pydantic & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# CrewAI SDK
crewai==0.1.25
langchain==0.0.350
langchain-community==0.0.1

# LLM Abstraction
litellm==1.11.1
openai==1.3.7
anthropic==0.7.7

# Docker SDK (for tool execution)
docker==6.1.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# WebSocket & Streaming
python-socketio==5.10.0
sse-starlette==1.8.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
ujson==5.9.0

# Monitoring & Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0

# Logging
structlog==23.2.0
